refactor(cli): extract reusable runtime from commands
