+++ cmd/otelop/start.go
@@ -2,48 +2,23 @@ package main
 
 import (
 	"context"
-	"encoding/base64"
-	"errors"
 	"fmt"
 	"io"
 	"log/slog"
-	"net"
-	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"sort"
 	"strings"
 	"syscall"
-	"time"
 
 	"github.com/urfave/cli/v3"
-	"go.opentelemetry.io/collector/otelcol"
-	"go.opentelemetry.io/contrib/bridges/otelslog"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/metric"
 
-	otelop "github.com/mashiro/otelop"
-	"github.com/mashiro/otelop/internal/broadcast"
-	"github.com/mashiro/otelop/internal/collector"
 	"github.com/mashiro/otelop/internal/config"
 	"github.com/mashiro/otelop/internal/daemon"
-	otelopexporter "github.com/mashiro/otelop/internal/exporter"
-	otelopgraphql "github.com/mashiro/otelop/internal/graphql"
-	"github.com/mashiro/otelop/internal/logger"
-	"github.com/mashiro/otelop/internal/selftelemetry"
-	"github.com/mashiro/otelop/internal/server"
-	"github.com/mashiro/otelop/internal/storage"
-	ws "github.com/mashiro/otelop/internal/websocket"
+	otelruntime "github.com/mashiro/otelop/internal/runtime"
 )
 
-func startCommand() *cli.Command {
-	// Load the TOML config file once at command-construction time. Its
-	// values become each flag's Default, so the resolved precedence is:
-	//   CLI flag > env var (Sources) > config file (Default) > built-in.
-	// A missing file is silently treated as "all defaults". A parse error
-	// is fatal at the next runtime call, surfaced from runStart.
+func startCommand(version string) *cli.Command {
 	cfg, cfgPath, cfgErr := config.Load()
 
 	return &cli.Command{
@@ -56,11 +31,7 @@ func startCommand() *cli.Command {
 			return nil, nil
 		},
 		Flags: []cli.Flag{
-			&cli.BoolFlag{
-				Name:    "foreground",
-				Aliases: []string{"f"},
-				Usage:   "run in the foreground instead of detaching",
-			},
+			&cli.BoolFlag{Name: "foreground", Aliases: []string{"f"}, Usage: "run in the foreground instead of detaching"},
 			&cli.StringFlag{Name: "http", Value: cfg.HTTPAddr, Usage: "Web UI + REST API listen address", Sources: cli.EnvVars("OTELOP_HTTP")},
 			&cli.StringFlag{Name: "otlp-grpc", Value: cfg.OTLPGRPCAddr, Usage: "OTLP gRPC receiver endpoint", Sources: cli.EnvVars("OTELOP_OTLP_GRPC")},
 			&cli.StringFlag{Name: "otlp-http", Value: cfg.OTLPHTTPAddr, Usage: "OTLP HTTP receiver endpoint", Sources: cli.EnvVars("OTELOP_OTLP_HTTP")},
@@ -79,74 +50,54 @@ func startCommand() *cli.Command {
 			&cli.StringFlag{Name: "log-level", Value: cfg.LogLevel, Usage: "log level (debug|info|warn|error)", Sources: cli.EnvVars("OTELOP_LOG_LEVEL")},
 			&cli.BoolFlag{Name: "debug", Value: cfg.Debug, Usage: "export otelop's own telemetry to itself", Sources: cli.EnvVars("OTELOP_DEBUG")},
 		},
-		Action:      runStart,
+		Action: func(ctx context.Context, cmd *cli.Command) error {
+			return runStart(ctx, cmd, version)
+		},
 		Description: fmt.Sprintf("Reads defaults from %s when present. Override with environment variables (OTELOP_HTTP, OTELOP_OTLP_GRPC, ...) or CLI flags.", cfgPath),
 	}
 }
 
-type proxyAuthOptions struct {
-	Type     string
-	Token    string
-	Username string
-	Password string
-	Headers  map[string]string
-}
-
 type startOptions struct {
-	HTTPAddr        string
-	OTLPGRPCAddr    string
-	OTLPHTTPAddr    string
-	AllowedHosts    []string
-	ProxyURL        string
-	ProxyProtocol   string
-	ProxyAuth       proxyAuthOptions
-	StoragePath     string
-	Retention       string
-	MaxSize         string
-	RenderWindowMax int
-	LogLevel        string
-	Debug           bool
-	Foreground      bool
+	runtime    otelruntime.Options
+	foreground bool
 }
 
-func optionsFromCmd(cmd *cli.Command) startOptions {
+func optionsFromCmd(cmd *cli.Command, version string) startOptions {
 	return startOptions{
-		HTTPAddr:      cmd.String("http"),
-		OTLPGRPCAddr:  cmd.String("otlp-grpc"),
-		OTLPHTTPAddr:  cmd.String("otlp-http"),
-		AllowedHosts:  splitCSV(cmd.String("allowed-hosts")),
-		ProxyURL:      strings.TrimSpace(cmd.String("proxy-url")),
-		ProxyProtocol: normalizeProxyProtocol(cmd.String("proxy-protocol")),
-		ProxyAuth: proxyAuthOptions{
-			Type:     normalizeProxyAuthType(cmd.String("proxy-auth-type")),
-			Token:    cmd.String("proxy-auth-token"),
-			Username: cmd.String("proxy-auth-username"),
-			Password: cmd.String("proxy-auth-password"),
-			Headers:  parseHeaderArgs(cmd.StringSlice("proxy-header")),
+		runtime: otelruntime.Options{
+			Version:         version,
+			HTTPAddr:        cmd.String("http"),
+			OTLPGRPCAddr:    cmd.String("otlp-grpc"),
+			OTLPHTTPAddr:    cmd.String("otlp-http"),
+			AllowedHosts:    splitCSV(cmd.String("allowed-hosts")),
+			ProxyURL:        strings.TrimSpace(cmd.String("proxy-url")),
+			ProxyProtocol:   strings.ToLower(strings.TrimSpace(cmd.String("proxy-protocol"))),
+			StoragePath:     strings.TrimSpace(cmd.String("storage-path")),
+			Retention:       cmd.String("retention"),
+			MaxSize:         cmd.String("max-size"),
+			RenderWindowMax: cmd.Int("render-window-max"),
+			LogLevel:        cmd.String("log-level"),
+			Debug:           cmd.Bool("debug"),
+			ProxyAuth: otelruntime.ProxyAuthOptions{
+				Type:     strings.ToLower(strings.TrimSpace(cmd.String("proxy-auth-type"))),
+				Token:    cmd.String("proxy-auth-token"),
+				Username: cmd.String("proxy-auth-username"),
+				Password: cmd.String("proxy-auth-password"),
+				Headers:  parseHeaderArgs(cmd.StringSlice("proxy-header")),
+			},
 		},
-		StoragePath:     strings.TrimSpace(cmd.String("storage-path")),
-		Retention:       cmd.String("retention"),
-		MaxSize:         cmd.String("max-size"),
-		RenderWindowMax: cmd.Int("render-window-max"),
-		LogLevel:        cmd.String("log-level"),
-		Debug:           cmd.Bool("debug"),
-		Foreground:      cmd.Bool("foreground"),
+		foreground: cmd.Bool("foreground"),
 	}
 }
 
-// splitCSV splits a comma-separated flag/env value (e.g. --allowed-hosts,
-// OTELOP_ALLOWED_HOSTS) into its trimmed, non-empty entries. Returns nil for
-// an empty/all-whitespace input so callers see the same "no entries" shape
-// TOML's native array gives for an omitted key.
 func splitCSV(v string) []string {
 	fields := strings.Split(v, ",")
 	out := make([]string, 0, len(fields))
 	for _, f := range fields {
 		f = strings.TrimSpace(f)
-		if f == "" {
-			continue
+		if f != "" {
+			out = append(out, f)
 		}
-		out = append(out, f)
 	}
 	if len(out) == 0 {
 		return nil
@@ -154,52 +105,41 @@ func splitCSV(v string) []string {
 	return out
 }
 
-func runStart(ctx context.Context, cmd *cli.Command) error {
-	opts := optionsFromCmd(cmd)
-	if err := validateProxyOptions(opts); err != nil {
+func runStart(ctx context.Context, cmd *cli.Command, version string) error {
+	opts := optionsFromCmd(cmd, version)
+	if err := otelruntime.Validate(opts.runtime); err != nil {
 		return err
 	}
-	if err := validateRenderWindowMax(opts.RenderWindowMax); err != nil {
-		return err
-	}
-	if !daemon.IsDaemonChild() && !opts.Foreground {
+	if !daemon.IsDaemonChild() && !opts.foreground {
 		return runDaemonParent(ctx)
 	}
-	return runServer(ctx, opts)
+	return runServer(ctx, opts.runtime)
 }
 
-// runServer runs the HTTP server, collector, and self-telemetry, then
-// blocks until SIGINT/SIGTERM. When invoked by the detached daemon child it
-// also persists metadata and signals the parent via the inherited pipe.
-func runServer(ctx context.Context, opts startOptions) error {
+func runServer(ctx context.Context, opts otelruntime.Options) error {
 	ready := daemon.ReadyPipe()
-	rt, err := bootstrap(ctx, opts)
+	rt, err := otelruntime.Start(ctx, opts)
 	if err != nil {
 		daemon.SignalError(ready, err)
 		return err
 	}
-	defer rt.shutdown()
+	defer rt.Shutdown()
 
 	if ready != nil {
 		meta := daemon.Metadata{
 			PID:           os.Getpid(),
-			StartedAt:     rt.startedAt,
+			StartedAt:     rt.StartedAt(),
 			HTTPAddr:      opts.HTTPAddr,
 			OTLPGRPCAddr:  opts.OTLPGRPCAddr,
 			OTLPHTTPAddr:  opts.OTLPHTTPAddr,
-			ProxyURL:      redactURL(opts.ProxyURL),
+			ProxyURL:      otelruntime.RedactURL(opts.ProxyURL),
 			ProxyProtocol: opts.ProxyProtocol,
-			Version:       version,
+			Version:       opts.Version,
 		}
 		if err := daemon.WriteMetadata(meta); err != nil {
 			daemon.SignalError(ready, err)
 			return err
 		}
-		// Acquire an advisory flock on the metadata file and hold the fd
-		// for the rest of the process lifetime. `otelop status`/`stop`
-		// probe this lock to distinguish a live daemon from stale
-		// metadata, which makes the check immune to PID recycling after a
-		// crash.
 		lockFile, err := daemon.LockMetadata()
 		if err != nil {
 			_ = daemon.RemoveState()
@@ -210,10 +150,10 @@ func runServer(ctx context.Context, opts startOptions) error {
 		defer func() { _ = daemon.RemoveState() }()
 		daemon.SignalReady(ready)
 	} else {
-		rt.printBanner(os.Stderr, opts)
+		printStartBanner(os.Stderr, opts)
 	}
 
-	rt.waitForSignal(ctx)
+	waitForShutdown(ctx, rt.Done())
 	return nil
 }
 
@@ -254,211 +194,12 @@ func runDaemonParent(ctx context.Context) error {
 	return nil
 }
 
-type runtime struct {
-	cancel            context.CancelFunc
-	startedAt         time.Time
-	storage           *storage.Storage
-	hub               *ws.Hub
-	srv               *server.Server
-	col               *otelcol.Collector
-	shutdownTelemetry func(context.Context) error
-}
-
-func bootstrap(ctx context.Context, opts startOptions) (*runtime, error) {
-	level, err := logger.ParseLevel(opts.LogLevel)
-	if err != nil {
-		return nil, err
-	}
-	logger.Setup(level)
-
-	ctx, cancel := context.WithCancel(ctx)
-
-	rt := &runtime{
-		cancel: cancel,
-		// Round(0) drops the monotonic reading so uptime spans system sleep.
-		startedAt: time.Now().Round(0),
-	}
-
-	rt.hub = ws.NewHub()
-	go rt.hub.Run(ctx)
-
-	storagePath, retention, maxSize, err := resolveStorageOptions(opts)
-	if err != nil {
-		rt.shutdown()
-		return nil, err
-	}
-
-	// st is declared before Open so the OnCommitBatch closure below can capture a
-	// reference to the very Storage it will be attached to (broadcast.New
-	// needs *storage.Storage to query back through). Safe because Open never
-	// invokes OnCommitBatch synchronously during construction — only later, from
-	// the writer goroutine, once st is already assigned.
-	var st *storage.Storage
-	onAdd := func(ctx context.Context, sig broadcast.SignalType, data any) {
-		rt.hub.BroadcastContext(ctx, ws.Message{Type: sig, Data: data})
-	}
-	st, err = storage.Open(ctx, storage.Options{
-		Path:      storagePath,
-		Retention: retention,
-		MaxSize:   maxSize,
-		OnCommitBatch: func(deliveries []storage.CommitDelivery) {
-			broadcast.NewBatch(st, onAdd, func() bool { return rt.hub.ClientCount() > 0 })(deliveries)
-		},
-	})
-	if err != nil {
-		rt.shutdown()
-		return nil, fmt.Errorf("open storage %s: %w", storagePath, err)
-	}
-	rt.storage = st
-
-	runtimeInfo := otelopgraphql.RuntimeInfo{
-		Version:          version,
-		StartedAt:        rt.startedAt,
-		HTTPAddr:         opts.HTTPAddr,
-		OTLPGRPCAddr:     opts.OTLPGRPCAddr,
-		OTLPHTTPAddr:     opts.OTLPHTTPAddr,
-		ProxyURL:         redactURL(opts.ProxyURL),
-		ProxyProtocol:    opts.ProxyProtocol,
-		Debug:            opts.Debug,
-		LogLevel:         opts.LogLevel,
-		Retention:        retention,
-		StoragePath:      storagePath,
-		RetentionDisplay: opts.Retention,
-		MaxSizeDisplay:   opts.MaxSize,
-		RenderWindowMax:  opts.RenderWindowMax,
-	}
-	rt.srv = server.New(rt.storage, rt.hub, otelop.FrontendFS(), runtimeInfo, opts.AllowedHosts)
-
-	// Eager listen so port conflicts surface before we signal ready.
-	if err := rt.srv.Listen(ctx); err != nil {
-		rt.shutdown()
-		return nil, fmt.Errorf("bind %s: %w", opts.HTTPAddr, err)
-	}
-	go func() {
-		if err := rt.srv.Start(ctx); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			slog.Error("HTTP server error", "error", err)
-			rt.cancel()
-		}
-	}()
-
-	selfTelemetryEndpoint := ""
-	if opts.Debug {
-		ep, err := resolveLoopback(opts.OTLPGRPCAddr)
-		if err != nil {
-			rt.shutdown()
-			return nil, fmt.Errorf("invalid otlp-grpc address: %w", err)
-		}
-		selfTelemetryEndpoint = ep
-	}
-
-	slog.Debug("starting collector", "grpc", opts.OTLPGRPCAddr, "http", opts.OTLPHTTPAddr)
-	col, err := collector.New(otelopexporter.NewFactory(rt.storage), collector.Config{
-		GRPCEndpoint:          opts.OTLPGRPCAddr,
-		HTTPEndpoint:          opts.OTLPHTTPAddr,
-		ProxyURL:              opts.ProxyURL,
-		ProxyProtocol:         opts.ProxyProtocol,
-		ProxyHeaders:          buildProxyHeaders(opts.ProxyAuth),
-		LogLevel:              opts.LogLevel,
-		SelfTelemetryEndpoint: selfTelemetryEndpoint,
-	})
-	if err != nil {
-		rt.shutdown()
-		return nil, fmt.Errorf("failed to create collector: %w", err)
-	}
-	rt.col = col
-
-	colErrCh := make(chan error, 1)
-	go func() {
-		if err := col.Run(ctx); err != nil {
-			colErrCh <- err
-		}
-		close(colErrCh)
-	}()
-
-	if err := waitCollectorReady(ctx, col, colErrCh); err != nil {
-		rt.shutdown()
-		return nil, err
-	}
-
-	// Once ready, nothing else reads colErrCh — without this, a post-ready
-	// Run failure (e.g. the OTLP listener dying) would go unnoticed and
-	// `otelop status` would keep reporting "running" while ingestion is
-	// silently dead. rt.cancel() unwinds waitForSignal, which triggers the
-	// deferred rt.shutdown() in runServer.
-	go watchCollectorErrors(colErrCh, rt.cancel, slog.Default())
-
-	if opts.Debug {
-		slog.Debug("starting self-telemetry", "endpoint", selfTelemetryEndpoint)
-		result, err := selftelemetry.Setup(ctx, selfTelemetryEndpoint)
-		if err != nil {
-			rt.shutdown()
-			return nil, fmt.Errorf("failed to setup self-telemetry: %w", err)
-		}
-		rt.shutdownTelemetry = result.Shutdown
-
-		otelHandler := otelslog.NewHandler("otelop", otelslog.WithLoggerProvider(result.LoggerProvider))
-		logger.Setup(level, otelHandler)
-
-		if err := registerMetrics(rt.storage, rt.hub); err != nil {
-			rt.shutdown()
-			return nil, fmt.Errorf("failed to register metrics: %w", err)
-		}
-	}
-
-	return rt, nil
-}
-
-// waitCollectorReady blocks until col reports StateRunning, an error is
-// observed on errCh, or the budget elapses. Replaces the old 500 ms blind
-// sleep — successful binds now return in ~10-50 ms.
-func waitCollectorReady(ctx context.Context, col *otelcol.Collector, errCh <-chan error) error {
-	const budget = 2 * time.Second
-	const tick = 10 * time.Millisecond
-	deadline := time.NewTimer(budget)
-	defer deadline.Stop()
-	ticker := time.NewTicker(tick)
-	defer ticker.Stop()
-	for {
-		select {
-		case err, ok := <-errCh:
-			if !ok {
-				return errors.New("collector exited before becoming ready")
-			}
-			if err != nil {
-				return fmt.Errorf("collector failed to start: %w", err)
-			}
-		case <-ticker.C:
-		case <-deadline.C:
-			return fmt.Errorf("collector did not become ready within %s", budget)
-		case <-ctx.Done():
-			return ctx.Err()
-		}
-		if col.GetState() == otelcol.StateRunning {
-			return nil
-		}
-	}
-}
-
-// watchCollectorErrors observes errCh for a collector failure that happens
-// after waitCollectorReady already returned success, and cancels the daemon
-// so the failure isn't silent. errCh is only ever populated when col.Run
-// returns a non-nil error; a clean shutdown (col.Shutdown() making Run
-// return nil) closes it without sending, which is a no-op here.
-func watchCollectorErrors(errCh <-chan error, cancel context.CancelFunc, logger *slog.Logger) {
-	err, ok := <-errCh
-	if !ok || err == nil {
-		return
-	}
-	logger.Error("collector stopped unexpectedly, shutting down", "error", err)
-	cancel()
-}
-
-func (r *runtime) printBanner(w io.Writer, opts startOptions) {
+func printStartBanner(w io.Writer, opts otelruntime.Options) {
 	suffix := ""
 	if opts.Debug {
 		suffix = " (debug)"
 	}
-	storagePath, err := resolveStoragePath(opts.StoragePath)
+	storagePath, err := otelruntime.ResolveStoragePath(opts.StoragePath)
 	if err != nil {
 		storagePath = opts.StoragePath
 	}
@@ -471,219 +212,16 @@ func (r *runtime) printBanner(w io.Writer, opts startOptions) {
 	})
 }
 
-// resolveStoragePath returns the DuckDB file location: raw if non-empty,
-// otherwise config.DefaultStoragePath()'s XDG-derived default.
-func resolveStoragePath(raw string) (string, error) {
-	if raw != "" {
-		return raw, nil
+func waitForShutdown(ctx context.Context, runtimeDone <-chan struct{}) {
+	sigCh := make(chan os.Signal, 1)
+	signal.Notify(sigCh, syscall.SIGINT, syscall.SIGTERM)
+	defer signal.Stop(sigCh)
+	select {
+	case <-sigCh:
+	case <-ctx.Done():
+	case <-runtimeDone:
 	}
-	return config.DefaultStoragePath()
-}
-
-// resolveStorageOptions turns the CLI-level startOptions.StoragePath/Retention/MaxSize
-// strings into the (path, retention, maxSize) triple internal/storage.Options
-// wants, resolving the XDG default path and creating its parent directory
-// (internal/storage.Open does not create directories itself) when no
-// explicit --storage-path was given.
-func resolveStorageOptions(opts startOptions) (path string, retention time.Duration, maxSize int64, err error) {
-	path, err = resolveStoragePath(opts.StoragePath)
-	if err != nil {
-		return "", 0, 0, fmt.Errorf("resolve storage path: %w", err)
-	}
-	if path != "" {
-		if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-			return "", 0, 0, fmt.Errorf("create storage directory: %w", err)
-		}
-	}
-	retention, err = config.ParseRetention(opts.Retention)
-	if err != nil {
-		return "", 0, 0, err
-	}
-	maxSize, err = config.ParseMaxSize(opts.MaxSize)
-	if err != nil {
-		return "", 0, 0, err
-	}
-	return path, retention, maxSize, nil
-}
-
-// validateRenderWindowMax rejects a non-positive --render-window-max/
-// OTELOP_RENDER_WINDOW_MAX value at startup: the frontend mounts exactly this
-// many <tr> rows at once in the traces/metrics/logs tables (no virtualization
-// library — frontend/src/hooks/use-render-window.ts), so anything less than 1
-// would leave those tables permanently empty.
-func validateRenderWindowMax(v int) error {
-	if v < 1 {
-		return fmt.Errorf("render-window-max must be >= 1, got %d", v)
-	}
-	return nil
-}
-
-func normalizeProxyProtocol(v string) string {
-	return strings.ToLower(strings.TrimSpace(v))
-}
-
-func normalizeProxyAuthType(v string) string {
-	return strings.ToLower(strings.TrimSpace(v))
-}
-
-func validateProxyOptions(opts startOptions) error {
-	if opts.ProxyURL == "" && opts.ProxyProtocol == "" {
-		if opts.ProxyAuth.Type != "" || hasProxyAuthFields(opts.ProxyAuth) {
-			return errors.New("proxy auth requires --proxy-url and --proxy-protocol")
-		}
-		return nil
-	}
-	if opts.ProxyURL == "" {
-		return errors.New("proxy-protocol requires --proxy-url")
-	}
-	if opts.ProxyProtocol == "" {
-		return errors.New("proxy-url requires --proxy-protocol (grpc|http)")
-	}
-	u, err := url.Parse(opts.ProxyURL)
-	if err != nil {
-		return fmt.Errorf("invalid proxy-url: %w", err)
-	}
-	if u.User != nil {
-		return errors.New("proxy-url must not contain embedded credentials; use --proxy-auth-* instead")
-	}
-	if err := validateProxyAuth(opts.ProxyAuth); err != nil {
-		return err
-	}
-	switch opts.ProxyProtocol {
-	case "grpc":
-		if err := validateGRPCProxyURL(u, opts.ProxyURL); err != nil {
-			return err
-		}
-		return validateNoSelfProxy(u, opts.ProxyURL, opts.OTLPGRPCAddr, "grpc")
-	case "http":
-		if err := validateHTTPProxyURL(u, opts.ProxyURL); err != nil {
-			return err
-		}
-		return validateNoSelfProxy(u, opts.ProxyURL, opts.OTLPHTTPAddr, "http")
-	default:
-		return fmt.Errorf("invalid proxy-protocol %q: want grpc or http", opts.ProxyProtocol)
-	}
-}
-
-func validateGRPCProxyURL(u *url.URL, raw string) error {
-	if u.Scheme == "" {
-		if raw == "" || strings.Contains(raw, "/") {
-			return fmt.Errorf("invalid proxy-url %q for grpc: want host:port or http(s)://host:port", raw)
-		}
-		return nil
-	}
-	switch strings.ToLower(u.Scheme) {
-	case "http", "https":
-		if u.Host == "" {
-			return fmt.Errorf("invalid proxy-url %q for grpc: missing host", raw)
-		}
-		return nil
-	default:
-		return fmt.Errorf("invalid proxy-url %q for grpc: unsupported scheme %q", raw, u.Scheme)
-	}
-}
-
-func validateHTTPProxyURL(u *url.URL, raw string) error {
-	if u.Scheme == "" || u.Host == "" {
-		return fmt.Errorf("invalid proxy-url %q for http: want http://host:port or https://host:port", raw)
-	}
-	switch strings.ToLower(u.Scheme) {
-	case "http", "https":
-		return nil
-	default:
-		return fmt.Errorf("invalid proxy-url %q for http: unsupported scheme %q", raw, u.Scheme)
-	}
-}
-
-func validateProxyAuth(auth proxyAuthOptions) error {
-	switch auth.Type {
-	case "":
-		if hasProxyAuthFields(auth) {
-			return errors.New("proxy auth fields require --proxy-auth-type")
-		}
-		return nil
-	case "bearer":
-		if auth.Token == "" {
-			return errors.New("proxy-auth-type bearer requires --proxy-auth-token")
-		}
-		if auth.Username != "" || auth.Password != "" || len(auth.Headers) > 0 {
-			return errors.New("proxy-auth-type bearer only supports --proxy-auth-token")
-		}
-	case "basic":
-		if auth.Username == "" || auth.Password == "" {
-			return errors.New("proxy-auth-type basic requires --proxy-auth-username and --proxy-auth-password")
-		}
-		if auth.Token != "" || len(auth.Headers) > 0 {
-			return errors.New("proxy-auth-type basic only supports username/password")
-		}
-	case "headers":
-		if len(auth.Headers) == 0 {
-			return errors.New("proxy-auth-type headers requires at least one --proxy-header")
-		}
-		if auth.Token != "" || auth.Username != "" || auth.Password != "" {
-			return errors.New("proxy-auth-type headers only supports --proxy-header")
-		}
-	default:
-		return fmt.Errorf("invalid proxy-auth-type %q: want bearer, basic, or headers", auth.Type)
-	}
-	return nil
-}
-
-func validateNoSelfProxy(u *url.URL, proxyURL, listenAddr, protocol string) error {
-	target, err := comparableProxyHostPort(u, proxyURL, protocol)
-	if err != nil {
-		return err
-	}
-	local, err := normalizeHostPort(listenAddr)
-	if err != nil {
-		return err
-	}
-	if target == local {
-		return fmt.Errorf("proxy-url %q points back to otelop's own OTLP %s listener %q", proxyURL, protocol, listenAddr)
-	}
-	return nil
-}
-
-func comparableProxyHostPort(u *url.URL, proxyURL, protocol string) (string, error) {
-	if protocol == "grpc" && u.Scheme == "" {
-		return normalizeHostPort(proxyURL)
-	}
-	return normalizeHostPort(u.Host)
-}
-
-func normalizeHostPort(addr string) (string, error) {
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		return "", err
-	}
-	host = strings.ToLower(strings.Trim(host, "[]"))
-	switch host {
-	case "", "0.0.0.0", "::", "::1", "127.0.0.1", "localhost":
-		host = "localhost"
-	}
-	return net.JoinHostPort(host, port), nil
-}
-
-func hasProxyAuthFields(auth proxyAuthOptions) bool {
-	return auth.Token != "" || auth.Username != "" || auth.Password != "" || len(auth.Headers) > 0
-}
-
-func buildProxyHeaders(auth proxyAuthOptions) map[string]string {
-	switch auth.Type {
-	case "bearer":
-		return map[string]string{"Authorization": "Bearer " + auth.Token}
-	case "basic":
-		token := base64.StdEncoding.EncodeToString([]byte(auth.Username + ":" + auth.Password))
-		return map[string]string{"Authorization": "Basic " + token}
-	case "headers":
-		out := make(map[string]string, len(auth.Headers))
-		for k, v := range auth.Headers {
-			out[k] = v
-		}
-		return out
-	default:
-		return nil
-	}
+	slog.Info("shutting down...")
 }
 
 func parseHeaderArgs(args []string) map[string]string {
@@ -692,19 +230,11 @@ func parseHeaderArgs(args []string) map[string]string {
 	}
 	out := make(map[string]string, len(args))
 	for _, arg := range args {
-		if arg == "" {
-			continue
-		}
 		k, v, ok := strings.Cut(arg, "=")
-		if !ok {
-			continue
-		}
 		k = strings.TrimSpace(k)
-		v = strings.TrimSpace(v)
-		if k == "" {
-			continue
+		if ok && k != "" {
+			out[k] = strings.TrimSpace(v)
 		}
-		out[k] = v
 	}
 	if len(out) == 0 {
 		return nil
@@ -713,9 +243,6 @@ func parseHeaderArgs(args []string) map[string]string {
 }
 
 func headerPairs(headers map[string]string) []string {
-	if len(headers) == 0 {
-		return nil
-	}
 	keys := make([]string, 0, len(headers))
 	for k := range headers {
 		keys = append(keys, k)
@@ -728,105 +255,9 @@ func headerPairs(headers map[string]string) []string {
 	return out
 }
 
-func redactURL(raw string) string {
-	if raw == "" {
-		return ""
-	}
-	u, err := url.Parse(raw)
-	if err != nil || u.User == nil {
-		return raw
-	}
-	u.User = url.UserPassword("REDACTED", "REDACTED")
-	return u.String()
-}
-
 func formatProxyStatus(proxyURL, proxyProtocol string) string {
 	if proxyURL == "" || proxyProtocol == "" {
 		return "disabled"
 	}
 	return fmt.Sprintf("%s %s", strings.ToUpper(proxyProtocol), proxyURL)
-}
-
-func (r *runtime) waitForSignal(ctx context.Context) {
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, syscall.SIGINT, syscall.SIGTERM)
-	select {
-	case <-sigCh:
-	case <-ctx.Done():
-	}
-	slog.Info("shutting down...")
-}
-
-func (r *runtime) shutdown() {
-	if r == nil {
-		return
-	}
-	shutdownCtx := context.Background()
-	if r.shutdownTelemetry != nil {
-		if err := r.shutdownTelemetry(shutdownCtx); err != nil {
-			slog.Error("self-telemetry shutdown error", "error", err)
-		}
-	}
-	if r.col != nil {
-		r.col.Shutdown()
-	}
-	if r.srv != nil {
-		if err := r.srv.Shutdown(shutdownCtx); err != nil {
-			slog.Error("HTTP server shutdown error", "error", err)
-		}
-	}
-	if r.storage != nil {
-		if err := r.storage.Close(); err != nil {
-			slog.Error("storage shutdown error", "error", err)
-		}
-	}
-	if r.cancel != nil {
-		r.cancel()
-	}
-}
-
-func registerMetrics(s *storage.Storage, hub *ws.Hub) error {
-	meter := otel.Meter("otelop")
-	if err := s.RegisterTelemetry(meter); err != nil {
-		return err
-	}
-
-	traceGauge, err := meter.Int64ObservableGauge("otelop.store.traces",
-		metric.WithDescription("Number of traces in the store"),
-	)
-	if err != nil {
-		return err
-	}
-	metricGauge, err := meter.Int64ObservableGauge("otelop.store.metrics",
-		metric.WithDescription("Number of metric series in the store"),
-	)
-	if err != nil {
-		return err
-	}
-	logGauge, err := meter.Int64ObservableGauge("otelop.store.logs",
-		metric.WithDescription("Number of log entries in the store"),
-	)
-	if err != nil {
-		return err
-	}
-	if _, err := meter.RegisterCallback(func(ctx context.Context, o metric.Observer) error {
-		traces, metrics, logs, err := s.Counts(ctx)
-		if err != nil {
-			return err
-		}
-		o.ObserveInt64(traceGauge, int64(traces))
-		o.ObserveInt64(metricGauge, int64(metrics))
-		o.ObserveInt64(logGauge, int64(logs))
-		return nil
-	}, traceGauge, metricGauge, logGauge); err != nil {
-		return err
-	}
-	_, err = meter.Int64ObservableGauge("otelop.websocket.clients",
-		metric.WithDescription("Number of connected WebSocket clients"),
-		metric.WithInt64Callback(func(_ context.Context, o metric.Int64Observer) error {
-			o.Observe(int64(hub.ClientCount()))
-			return nil
-		}),
-	)
-	return err
 }
