+++ internal/netutil/address.go
@@ -1,0 +1,15 @@
+package netutil
+
+import "net"
+
+// Loopback converts a listen address into an address a local client can use.
+func Loopback(listenAddr string) (string, error) {
+	host, port, err := net.SplitHostPort(listenAddr)
+	if err != nil {
+		return "", err
+	}
+	if host == "" || host == "0.0.0.0" || host == "::" {
+		host = "localhost"
+	}
+	return host + ":" + port, nil
+}
